Move ReferenceKind
